% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrtr-package.R
\docType{package}
\name{discrtr-package}
\alias{discrtr}
\alias{discrtr-package}
\title{discrtr: A companion package for the book Introduction to Discrete Choice Analysis with `R`}
\description{
This package includes data files and supporting materials for the book.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paezha/discrtr}
  \item Report bugs at \url{https://github.com/paezha/discrtr/issues}
}

}
\author{
\strong{Maintainer}: Antonio Paez \email{paezha@mcmaster.ca} (\href{https://orcid.org/0000-0001-6912-9919}{ORCID})

Authors:
\itemize{
  \item Genevieve Boisjoly \email{gboisjoly@polymtl.ca} (\href{https://orcid.org/0000-0001-5375-7750}{ORCID})
}

}
\keyword{internal}
