% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrtr-package.R
\name{discrtr_template}
\alias{discrtr_template}
\title{discrtr template.}
\format{
An R Markdown file
}
\usage{
discrtr_template()
}
\description{
Template for creating a computational notebook for use with Discrete Choice Analysis with R
}
\keyword{template}
