% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrtr-package.R
\docType{data}
\name{urban_types}
\alias{urban_types}
\title{Hamilton City boundaries and urban types}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 3 rows and 2 columns.
}
\usage{
data(urban_types)
}
\description{
Classification of regions by type: urban, suburban, and rural. This is based on the rural boundaries of the city, and the planning regions.
}
\details{
#' @format A simple features data frame with 2 rows and 2 variables:
\describe{
\item{Type}{Classification of the region: "Urban", "Suburban", "Rural"}
\item{geometry}{Geometry of polygons}
}
}
\keyword{datasets}
\keyword{transportation}
