% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrtr-package.R
\docType{package}
\name{discrtr-package}
\alias{discrtr}
\alias{discrtr-package}
\title{discrtr: A Companion Package for the Book "Discrete Choice Analysis with 'R'"}
\description{
Templates and data files to support "Discrete Choice Analysis with R", Páez, A. and Boisjoly, G. (2023) \doi{10.1007/978-3-031-20719-8}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paezha/discrtr}
  \item Report bugs at \url{https://github.com/paezha/discrtr/issues}
}

}
\author{
\strong{Maintainer}: Antonio Paez \email{paezha@mcmaster.ca} (\href{https://orcid.org/0000-0001-6912-9919}{ORCID})

Authors:
\itemize{
  \item Genevieve Boisjoly \email{gboisjoly@polymtl.ca} (\href{https://orcid.org/0000-0001-5375-7750}{ORCID})
}

}
\keyword{internal}
