% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrtr-package.R
\docType{package}
\name{discrtr-package}
\alias{discrtr}
\alias{discrtr-package}
\title{discrtr: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paezha/discrtr}
  \item Report bugs at \url{https://github.com/paezha/discrtr/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
